requirements.txt
allure-behave==2.13.2
allure-python-commons==2.13.2
attrs==25.3.0
behave==1.2.6
certifi==2025.7.14
charset-normalizer==3.4.2
h11==0.16.0
idna==3.10
outcome==1.3.0.post0
parse==1.20.2
parse_type==0.6.4
pluggy==1.6.0
PySocks==1.7.1
python-dateutil==2.9.0
pytz==2025.2
requests==2.32.3
selenium==4.17.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
urllib3==2.5.0
wsproto==1.2.0
